Made the UI for the player health, visual feed back for taking damage and healing
Made the HealthOrb, HealthBar and ManaIndicator components, made it so it reacts to healing and taking damage properly, didn't make it flexible to add or take away health, it's not flexible in that regard. Next step is to add mana to the bar every time the player hits an enemy and then take away mana when healing. That's the next step, when I get to it.