Made the pickaxe have a path and a path follower and a projectile firing component that drives that code
It's a little dirty in some places and could be touched up, but it works, all the pickaxe needs is a position and it will arc int that direction, the speed of the travel is controlled from the pickaxe itself. The driver code can be xpanded to shoot any kind of projectile. The functuions should be cleaned up to make it more clear what is happening.

The pickaxe onlz makes sure to progress the path follow, the path itself and the target is defined from the driver code.